requirements.txt
discord.py==2.0.1
discord==1.7.3
#discord-py-slash-command==3.0.3
mysql_connector_repackaged==0.3.1
mysql-connector-python==8.0.30
python-dotenv==0.19.2
requests==2.26.0
markdownify==0.10.1
python-version==0.0.2
websockets==10.1
emoji==1.6.3
sentry-sdk==1.9.0